jun: publish "api.on" for domain proxy service
api.on and the proxy domain service are running on the same host. Thus
the "api.on" domain needs to be published explicitly, since the default
webserver on jun.on runs on port 81 and is only available via the domain
proxy service.

We may remove this announcement, when there is a newer firmware (0.5.5
or higher). For now the latest published firmware (0.5.4) uses
"http://api.on/" for its traceroute and map visualization.