ondataservice_firmware_versions: separate graphs for all firmware versions
The main graph now shows a distribution of the installed firmware
versions (instead of absolute numbers).
The sub-graphs show the absolute numbers and mention all accesspoints
with this version in their "extinfo" field below the graphs.

The new graphs use a different base name (since the new values are
incompatible). Thus the old graphs are gone.