olsr2: create generic loopback systemd-networkd config
We need to prepare a generic systemd-networkd interface configuration attached to the loopback. This will be reused for other roles (e.g. gateway-server) to add additional IPs to the same loopback interface.

Therefor only create the file if it not exists and "just" make sure that the proper ULA IPv6 is set.