gateway-server: update the loopback systemd-networkd config
We check if the systemd-networkd configuration exists and will create it otherwise. Also make sure the needed Opennet OLSRv1 mesh ip is added to the configuration.

This removes the separate systemd-networkd files for the olsr2 and gateway-server role and use one file. Otherwise only the last config will be read by networkd and we will miss either IPv4 oder IPv6 loopback ip.

Since we create always a file of no file exists there are no mutual requirements or order need to apply the olsr2 and gateway-server rule in regards to the systemd-networkd configuration.