buildbot-controller: move webinterface to dev
Recently we introduces buildbot as replacement for the old Opennet DEV environment based on trac. During the change we maintained both services in parallel with "dev2" and "dev".

Using not again only "dev" as name - mainly for the webinterface of the buildbot worker status. Hosting is been done on server "goat".