Enable switch for building only small amount of archs
Sadly, the code was not working as expected. Requesting the value
of `force_make_arch_all` did not always returned the correct result.
I do not know how to fix this therefor we reduce/simplify it.
In case somebody needs fast builds for a limited amount of time
it can be hard coded here.