olsr: Allow applying olsr/olsr2 role alone
We want to be able to apply only the olsr/olsr2 without executing all other roles.
With tags we can choose on playbook start to execute only this role.
Some variables need to be declared globally because olsr/olsr2 is using these vars. When they are not declared globally the role will fail because gateway-server role is not included.